+++ src/runtime/stack.rs
@@ -181,19 +181,19 @@ pub struct StackFrame {
   // Line/col for debug info
   debug_pos: (usize, usize),
   // Source name for debug info
-  debug_source: Option<RantString>,
+  origin: Rc<RantString>,
 }
 
 impl StackFrame {
   pub fn new(sequence: Rc<Sequence>, has_output: bool, prev_output: Option<&OutputWriter>) -> Self {
     Self {
+      origin: Rc::clone(&sequence.origin),
       sequence,
       output: if has_output { Some(OutputWriter::new(prev_output)) } else { None },
       started: false,
       pc: 0,
       intents: Default::default(),
       debug_pos: (0, 0),
-      debug_source: None,
     }
   }
 }
@@ -261,8 +261,6 @@ impl StackFrame {
   pub fn set_debug_info(&mut self, info: &DebugInfo) {
     match info {
       DebugInfo::Location { line, col } => self.debug_pos = (*line, *col),
-      DebugInfo::SourceName(name) => self.debug_source = Some(name.clone()),
-      DebugInfo::ScopeName(_) => todo!(),
     }
   }
 }
@@ -306,7 +304,7 @@ impl StackFrame {
 impl Display for StackFrame {
   fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
     write!(f, "[{}:{}:{}] in {}", 
-      self.debug_source.as_ref().map(|src| src.as_str()).unwrap_or("program"), 
+      self.origin.as_ref(), 
       self.debug_pos.0, 
       self.debug_pos.1,
       self.sequence.name().map(|name| name.as_str()).unwrap_or("???"),